backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
docker==7.1.0
PyGithub==2.5.0
websockets==14.1
python-multipart==0.0.18
redis==5.2.1
celery==5.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
boto3==1.35.86
stripe==11.4.1
tenacity==9.0.0
structlog==24.4.0
greenlet>=3.0.0
